+++ CHANGES.rst
@@ -184,3 +184,5 @@ Miscellaneous
 
 - Versioning of signature module for protocol specific support and future
   extensibility.
+
+- Removed binaries of Kathmandu.